Updated README.md for clarity on the fix process
Updated the docs to explain the apply_haproxy_config.sh & apply_mongo_config.sh steps better, and that they are an alternative to just starting with the fixed docker-compose-done.yml file.  Might need more explanation as why we start broken, to show you can apply a config to a running supervisor ring?